+++ src/main/java/learn/thyme/thyteams/ThymeleafApplicationConfiguration.java
@@ -2,8 +2,10 @@ package learn.thyme.thyteams;
 
 import io.github.wimdeblauwe.jpearl.InMemoryUniqueIdGenerator;
 import io.github.wimdeblauwe.jpearl.UniqueIdGenerator;
+import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
 import org.thymeleaf.templateresolver.ITemplateResolver;
 
@@ -23,5 +25,12 @@ public class ThymeleafApplicationConfiguration {
     @Bean
     public UniqueIdGenerator<UUID> uniqueIdGenerator() {
         return new InMemoryUniqueIdGenerator();
+    }
+
+    @Bean
+    public LocalValidatorFactoryBean localValidatorFactoryBean(MessageSource messageSource) {
+        LocalValidatorFactoryBean bean = new LocalValidatorFactoryBean();
+        bean.setValidationMessageSource(messageSource);
+        return bean;
     }
 }