+++ src/main/java/learn/thyme/thyteams/infrastructure/validation/ValidationGroupOne.java
@@ -1,0 +1,4 @@
+package learn.thyme.thyteams.infrastructure.validation;
+
+public interface ValidationGroupOne {
+}