+++ src/test/java/learn/thyme/thyteams/ThyteamsApplicationTests.java
@@ -1,9 +1,13 @@
 package learn.thyme.thyteams;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("spring-boot-test")
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 class ThyteamsApplicationTests {
 
     @Test